move logging configuration to core module
